+++ src/main/java/com/tsp/new_tsp_admin/api/domain/model/AdminModelEntity.java
@@ -38,7 +38,7 @@ public class AdminModelEntity extends NewCommonMappedClass {
     @Id
     @GeneratedValue(strategy = IDENTITY)
     @Column(name = "idx")
-    private Integer idx;
+    private Long idx;
 
     @Column(name = "category_cd")
     @Range(min = 1, max = 3, message = "모델 categoryCd는 1~3 사이 값만 입력할 수 있습니다.")
@@ -106,7 +106,7 @@ public class AdminModelEntity extends NewCommonMappedClass {
     private Integer viewCount;
 
     @Column(name = "agency_idx")
-    private Integer agencyIdx;
+    private Long agencyIdx;
 
     @Column(name = "career_list")
     @Convert(converter = CustomConverter.class)
